Fix: 프리셋 MQTT 발행 로직 수정 및 기타 디버깅

See merge request s13-webmobile3-sub1/S13P11B101!92

+++ backend/siseon-backend/src/main/java/siseon/backend/domain/main/Profile.java
@@ -37,7 +37,7 @@ public class Profile {
 
     @NotNull
     @Column(nullable = false)
-    private Float height;
+    private Integer height;
 
     @NotNull
     @Column(name = "left_vision", nullable = false)
